MiCarls: SFBUG:3287800 - on uninstall, uninstall addons before everything else.          Continue uninstalling bundles to be upgraded, after everything else.
A-BobArn: Make WixLocalization/@Culture optional (defaulting to neutral).

MiCarls: SFBUG:3285054 - Fix WIX_SUITE_EMBEDDEDNT property

RobMen: Support slipstream MSPs.

RobMen: Add VsixPackage install support to WixVSExtension.

WixBuild: Version 3.6.1629.0

PMarcu: SFBUG:3205933 - Persist relative download urls in the burn manifest.
        Fix Dark to not assume all payloads in a bundle are embedded.

MiCarls: SFBUG:3287802 - allow related bundles (and your own bundle) to be
         identified with Bundle/@Tag

RobMen: SFBUG:3292845 - enable parallel cache via Bundle/Chain/ParallelCache
        attribute.

RobMen: SFBUG:3172051 - fix documentation to correctly reference install
        location for WiX targets.

RobMen: SFBUG:3283837 - fix Votive About Box and Bundle template to correct
        version.
        SFBUG:3116741 - remove dead sln file.

HeathS: Added dependency/ref-counting support to Burn.

WixBuild: Version 3.6.1706.0

EricStJ: Adding support for netfx progress structure to communicate with .NET
         framework setup.

WixBuild: Version 3.6.1714.0

PMarcu: SFBUG:2981229 Removing LayoutDirectory and LayoutFile.

BMurri: Missing MSI 5.0 Support + DTF Bugs, including SFBUG:3072105

HeathS: Remove all burn dependents when a bundle is uninstalled.
        Remove orphaned registry keys from dependency graph for bundle.

A-BobArn: SFBUG:3293975: store state when suspending

BMurri: AdminExecuteSequence not processed correctly when building patches

A-BobArn: Add WIX_COMPRESSION_LEVEL environment variable detection to override
          authored or default compression level for cabinets.
          WIX_COMPRESSION_LEVEL must be empty or one of the same values as
          Media/@CompressionLevel: none, mszip, low, medium, high

EricStJ: SFBUG:3267882 Need to hide property for HideTarget CustomActions

MiCarls: SFBUG:3292534 Don't fail if a payload we're copying to is read-only.

EricStJ: SFBUG:3274080 Don't fail if CA dll is missing version

EricStJ: SFBUG:3155753 Support LogVisits using dontLog dir prop.
         SFBUG:3291011 Set cache mode when setting HttpExpires and
                       CacheControlMaxAge

BMurri: Prevents some instances of reading/passing unneeded data
        Fixes a race condition that causes successful CA calls to return
        failure to MSI

EricStJ: SFBUG:3109682 Don't log errors when doing a website search

HeathS: Changed default bundle provider key generation to use bundle ID.

RobMen: SFBUG:3303380 - create separate pipes for cache and apply threads to
        prevent collisions in messages coming back at the same time.

WixBuild: Version 3.6.1727.0

CAraman: fix dutil annotations

EricStJ: SFBUG:3135376 - failure to create websvcextension when exists with
         string-different but path-identical extension path.

RobMen: Do not force all packages using the WixDependencyExtension to codepage
        1252.

WixBuild: Version 3.6.1803.0

PMarcu: SFBUG:3307070 Generic fix to catch duplicate symbol issues early and
                      cover all of them that we aren't special handling.

BMurri: SFBUG: 3266592: Delta switch failure when an updated file is shared
                        across cabinets.

EricStJ: SFBUG: 3157348 Don't block uninstall if WebDir's web is missing.

MiCarls: SFBUG:3103010 & SFBUG:3102971 - make wix schema tolerate preprocessor
         variables being used in more places.

PMarcu: SFBUG: 3309263: Fix to create fragments for each component.

Jbae: Made ProjectDefineConstants properties static because they don't have to
      be dynamic

BMurri: SFBUG: 3175345: Stop dropping odd build output directories all over the
                        build tree.
        Also fixes doubling of BindPaths sent to light from msbuild.

EricStJ: SFBUG: 3181941 Make ManagedPipelineMode and ManagedRuntimeVersion
                        formatted

HeathS: Change default provider key generation for consistent SxS experience.
        SFBUG:3288070 - Fixed documentation for dependency extension.

ZhenyaS: SFBUG:3163741 Allow bundle authors to customize prerequisites UX
                       placeholders.
         SFBUG 3308501 Don't show source resolution in silent mode, but try
                       downloading netfx package instead
         SFBUG 3184248 Allow netfx to be installed in non-install scenarios
                       (e.g. layout, uninstall, repair, modify)
         SFBUG 3158640 When not finding a registry value during a search,
                       clear out a possibly stale variable. This allows us to
                       correctly detect that netfx is absent in uninstall,
                       repair, modify scenarios.

JBae: Fixed bug for unused TargetPdbName property.

WixBuild: Version 3.6.1811.0

--HG--
branch : wix36

+++ src/DTF/Libraries/WindowsInstaller/Installer.cs
@@ -747,7 +747,7 @@ public static partial class Installer
         }
         else
         {
-            ret = NativeMethods.MsiDeterminePatchSequence(product, userSid, (int) context, (uint) sequenceData.Length, sequenceData);
+            ret = NativeMethods.MsiDeterminePatchSequence(product, userSid, context, (uint) sequenceData.Length, sequenceData);
         }
 
         if (errorHandler != null)
